+++ lib/core/src/display/shape.rs
@@ -1,6 +1,5 @@
 //! This is a root module for shapes, 2-dimensional graphical elements.
 
 pub mod compound;
-pub mod glyph;
 pub mod primitive;
 pub mod text;
