+++ lib/core/src/display/shape/text/glyph.rs
