+++ lib/core/src/display/shape/text/glyph/glyph.glsl
