+++ lib/core/src/display/shape/text/glyph/msdf.rs
