+++ lib/core/src/display/shape/text/glyph/pen.rs
@@ -4,7 +4,7 @@
 
 use crate::prelude::*;
 
-use crate::display::shape::glyph::font::FontRenderInfo;
+use crate::display::shape::text::glyph::font::FontRenderInfo;
 
 use nalgebra::Vector2;
 
@@ -76,7 +76,7 @@ where CharIterator : Iterator<Item=char> {
 mod tests {
     use super::*;
 
-    use crate::display::shape::glyph::font::GlyphRenderInfo;
+    use crate::display::shape::text::glyph::font::GlyphRenderInfo;
 
     use basegl_core_msdf_sys::test_utils::TestAfterInit;
     use std::future::Future;
