+++ lib/core/src/display/shape/text/text_field/buffer/glyph_square.rs
